refactor: modularize Family
